Reorganize to clarify transport_inner as internal analog
