+++ src/rpc_correlation.rs
@@ -1,3 +1,16 @@
+/********************************************************************************
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
 use crate::{ClientId, ReqId, RequestId, SessionId};
 use lazy_static::lazy_static;
 use log::trace;
@@ -90,3 +103,5 @@ pub(crate) async fn remove_me_request_correlation(req_id: &ReqId) -> Result<Requ
 
     Ok(request_id)
 }
+
+// TODO: Add unit tests
