+++ src/vsomeip_offered_requested.rs
@@ -1,3 +1,16 @@
+/********************************************************************************
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
 use crate::{EventId, InstanceId, MethodId, ServiceId};
 use lazy_static::lazy_static;
 use std::collections::HashSet;
@@ -85,3 +98,5 @@ pub(crate) async fn insert_event_requested(
     let mut requested_events = REQUESTED_EVENTS.write().await;
     requested_events.insert((service_id, instance_id, event_id));
 }
+
+// TODO: Add unit tests
